# feat(query): optimize count(Nullable(col))