--- a/src/query/sql/src/planner/binder/table.rs
+++ b/src/query/sql/src/planner/binder/table.rs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 use std::collections::HashMap;
+use std::collections::HashSet;
 use std::sync::Arc;
 
 use common_ast::ast::Indirection;
@@ -50,6 +51,7 @@ use crate::plans::ConstantExpr;
 use crate::plans::LogicalGet;
 use crate::plans::Scalar;
 use crate::BindContext;
+use crate::ColumnSet;
 use crate::IndexType;
 
 impl<'a> Binder {
@@ -302,6 +304,9 @@ impl<'a> Binder {
         let mut bind_context = BindContext::with_parent(Box::new(bind_context.clone()));
         let columns = self.metadata.read().columns_by_table_index(table_index);
         let table = self.metadata.read().table(table_index).clone();
+
+        let mut column_indexs: ColumnSet = HashSet::new();
+        let mut has_path_indices = false;
         let mut col_stats: HashMap<IndexType, Option<ColumnStatistics>> = HashMap::new();
         for column in columns.iter() {
             let column_binding = ColumnBinding {
@@ -317,16 +322,22 @@ impl<'a> Binder {
                 },
             };
             bind_context.add_column_binding(column_binding);
-            let column_id = column.index();
+            column_indexs.insert(column.index());
+            if !has_path_indices {
+                // create table t (c0 tuple(int, int), c1 string null) in planner col index is : (0: c0, 1: c1, 2: c0:2 3: c0:2)
+                // in storage column (0: PrimitiveColumn, 1: PrimitiveColumn, 2: NullableColumn)
+                has_path_indices = column.has_path_indices();
+            }
             let col_stat = table
                 .table()
                 .column_statistics_provider()
                 .await?
-                .column_statistics(column_id as ColumnId);
-            col_stats.insert(column_id, col_stat);
+                .column_statistics(column.index() as ColumnId);
+            col_stats.insert(column.index(), col_stat);
         }
-        let stat = table.table().table_statistics()?;
 
+        let is_accurate = !has_path_indices && table.table().engine().to_lowercase() == "fuse";
+        let stat = table.table().table_statistics()?;
         Ok((
             SExpr::create_leaf(
                 LogicalGet {
@@ -338,6 +349,7 @@ impl<'a> Binder {
                     statistics: stat,
                     prewhere: None,
                     col_stats,
+                    is_accurate,
                 }
                 .into(),
             ),