--- a/src/query/sql/src/planner/optimizer/heuristic/decorrelate.rs
+++ b/src/query/sql/src/planner/optimizer/heuristic/decorrelate.rs
@@ -412,6 +412,7 @@ impl SubqueryRewriter {
                     statistics: None,
                     col_stats: HashMap::new(),
                     prewhere: None,
+                    is_accurate: false,
                 }
                 .into(),
             );