--- a/src/query/sql/src/planner/optimizer/property/mod.rs
+++ b/src/query/sql/src/planner/optimizer/property/mod.rs
@@ -20,6 +20,8 @@ mod property;
 mod stat;
 
 pub use builder::RelExpr;
+pub use column_stat::ColumnStat;
+pub use column_stat::ColumnStatSet;
 pub use enforcer::require_property;
 pub use property::ColumnSet;
 pub use property::Distribution;