--- a/src/query/sql/src/planner/optimizer/property/mod.rs
+++ b/src/query/sql/src/planner/optimizer/property/mod.rs
@@ -28,3 +28,4 @@ pub use property::Distribution;
 pub use property::PhysicalProperty;
 pub use property::RelationalProperty;
 pub use property::RequiredProperty;
+pub use property::Statistics;