--- a/src/query/sql/src/planner/optimizer/property/property.rs
+++ b/src/query/sql/src/planner/optimizer/property/property.rs
@@ -51,6 +51,8 @@ pub struct RelationalProperty {
 
     /// Statistics of columns, column index -> column stat
     pub column_stats: ColumnStatSet,
+    /// Statistics info is accurate
+    pub is_accurate: bool,
 }
 
 #[derive(Default, Clone)]