--- a/src/query/sql/src/planner/plans/aggregate.rs
+++ b/src/query/sql/src/planner/plans/aggregate.rs
@@ -20,6 +20,7 @@ use crate::optimizer::PhysicalProperty;
 use crate::optimizer::RelExpr;
 use crate::optimizer::RelationalProperty;
 use crate::optimizer::RequiredProperty;
+use crate::optimizer::Statistics;
 use crate::plans::LogicalOperator;
 use crate::plans::Operator;
 use crate::plans::PhysicalOperator;
@@ -154,16 +155,19 @@ impl LogicalOperator for Aggregate {
         // Derive used columns
         let mut used_columns = self.used_columns()?;
         used_columns.extend(input_prop.used_columns);
+        let column_stats = input_prop.statistics.column_stats;
+        let is_accurate = input_prop.statistics.is_accurate;
 
         Ok(RelationalProperty {
             output_columns,
             outer_columns,
             used_columns,
             cardinality,
-            precise_cardinality,
-
-            column_stats: input_prop.column_stats,
-            is_accurate: input_prop.is_accurate,
+            statistics: Statistics {
+                precise_cardinality,
+                column_stats,
+                is_accurate,
+            },
         })
     }