--- a/src/query/sql/src/planner/plans/eval_scalar.rs
+++ b/src/query/sql/src/planner/plans/eval_scalar.rs
@@ -114,6 +114,7 @@ impl LogicalOperator for EvalScalar {
             precise_cardinality,
 
             column_stats: Default::default(),
+            is_accurate: input_prop.is_accurate,
         })
     }