--- a/src/query/sql/src/planner/plans/filter.rs
+++ b/src/query/sql/src/planner/plans/filter.rs
@@ -19,6 +19,7 @@ use crate::optimizer::PhysicalProperty;
 use crate::optimizer::RelExpr;
 use crate::optimizer::RelationalProperty;
 use crate::optimizer::RequiredProperty;
+use crate::optimizer::Statistics;
 use crate::plans::LogicalOperator;
 use crate::plans::Operator;
 use crate::plans::PhysicalOperator;
@@ -102,10 +103,11 @@ impl LogicalOperator for Filter {
             cardinality,
             // TODO(leiysky): if the predicate is always true, then we can pass through
             // precise cardinality
-            precise_cardinality: None,
-
-            column_stats: Default::default(),
-            is_accurate: false,
+            statistics: Statistics {
+                precise_cardinality: None,
+                column_stats: Default::default(),
+                is_accurate: false,
+            },
         })
     }