--- a/src/query/sql/src/planner/plans/logical_join.rs
+++ b/src/query/sql/src/planner/plans/logical_join.rs
@@ -224,6 +224,7 @@ impl LogicalOperator for LogicalInnerJoin {
             precise_cardinality: None,
 
             column_stats: Default::default(),
+            is_accurate: false,
         })
     }