--- a/src/query/sql/src/planner/plans/logical_join.rs
+++ b/src/query/sql/src/planner/plans/logical_join.rs
@@ -21,6 +21,7 @@ use super::ScalarExpr;
 use crate::optimizer::ColumnSet;
 use crate::optimizer::RelExpr;
 use crate::optimizer::RelationalProperty;
+use crate::optimizer::Statistics;
 use crate::plans::LogicalOperator;
 use crate::plans::Operator;
 use crate::plans::PhysicalOperator;
@@ -221,10 +222,11 @@ impl LogicalOperator for LogicalInnerJoin {
             outer_columns,
             used_columns,
             cardinality,
-            precise_cardinality: None,
-
-            column_stats: Default::default(),
-            is_accurate: false,
+            statistics: Statistics {
+                precise_cardinality: None,
+                column_stats: Default::default(),
+                is_accurate: false,
+            },
         })
     }