… tests/logictest/suites/base/03_dml/03_0032_select_null_count