I hereby agree to the terms of the CLA available at: https://databend.rs/dev/policies/cla/

## Summary

If simple count(nullable(col)), fold it.

|now|before optimize|
|:--:|:--:|
| 0.002s | 0.091s |
 
```sql
CREATE TABLE orders
(
    o_orderkey       BIGINT  null,
    o_custkey        BIGINT  null,
    o_orderstatus    STRING  null,
    o_totalprice     DOUBLE  null,
    o_orderdate      DATE  null,
    o_orderpriority  STRING  null,  
    o_clerk          STRING  null, 
    o_shippriority   INTEGER  null,
    o_comment        STRING  null
);

mysql> select count(o_clerk) from orders;
+----------------+
| count(o_clerk) |
+----------------+
|       15000000 |
+----------------+
1 row in set (0.02 sec)
Read 1 rows, 1.00 B in 0.002 sec., 469.46 rows/sec., 469.46 B/sec.

mysql> select count(o_clerk) from orders;
+----------------+
| count(o_clerk) |
+----------------+
|       15000000 |
+----------------+
1 row in set (0.10 sec)
Read 15000000 rows, 330.81 MiB in 0.091 sec., 164.49 million rows/sec., 3.54 GiB/sec.

```

Closes #8804