@dantengsky 2022-11-15
> Ping @dantengsky -- Can the statistical information used in this PR be guaranteed to be 100% accurate?

the `number_of_distinct_values` provided currently, is FAKED, just the number of rows; @lichuang is working on a real impl of ndv: https://github.com/datafuselabs/databend/pull/8521, but it will only provide approximated values.

the `null_count` shall be accurate, though