@TCeason 2022-11-16
> the column_stats of hive in planning stat is not correct. may leading wrong result in hive sql?

Could you please give me a demo, I can test it.