@sundy-li 2022-11-16
> > Could you please give me a demo, I can test it.
> 
> hive doesn't provide column_statistics_provider now, even provided most hive tables on-product does't have columns stats, and columns stats is not 100% trustable.

That's ok. 

HiveTable return None TableStatistics
```
 fn table_statistics(&self) -> Result<Option<TableStatistics>> {
        Ok(None)
    }
```