@sundy-li 2022-11-16
That' ok, we can introduce another field named `is_accruate` to indicate whether this `TableStatistics ` is fully trusted.