@TCeason 2022-11-16
> That' ok, we can introduce another field named `is_accruate` to indicate whether this `TableStatistics ` is fully trusted.

I add a function in src/query/catalog/src/table.rs, now only fuse table is accurate.

```rust
fn is_accurate(&self) -> bool {
    false
}
```